# services/contract/ledger-openapi.yaml
openapi: 3.0.3

info:
  title: Trade Lab API — Ledger
  version: 1.0.0
  description: Account management endpoints

servers:
  - url: http://localhost:8080/api/v1
    description: Local development

tags:
  - name: Ledger
    description: Account opening and listing

paths:

  /accounts:
    post:
      tags:
        - Ledger
      summary: Open a new account
      operationId: openAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenAccountRequest'
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          description: Validation failure — missing required field or invalid currency
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: 400
                error: Validation failed
                details:
                  - "currency: must be one of USD, GBP, EUR"
        '401':
          description: Unauthorised
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                status: 403
                error: User not found
                details:
                  - "No user found with id: 550e8400-e29b-41d4-a716-446655440000"

    get:
      tags:
        - Ledger
      summary: List accounts for a user
      operationId: listAccounts
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [ACTIVE, SUSPENDED, CLOSED]
      responses:
        '200':
          description: List of accounts for the user (may be empty)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountListResponse'
        '401':
          description: Unauthorised
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /accounts/{accountId}/top-up:
    post:
      tags:
        - Ledger
      summary: Top up an account balance
      operationId: topUpAccount
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TopUpAccountRequest'
      responses:
        '200':
          description: Top-up successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopUpAccountResponse'
        '400':
          description: Validation failure — invalid amount
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorised
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Account not available for top-up
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /accounts/{accountId}/transactions:
    get:
      tags:
        - Ledger
      summary: List transactions for an account
      operationId: getAccountTransactions
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: userId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: Paginated list of transactions for the account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionListResponse'
        '401':
          description: Unauthorised
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Account not owned by the requesting user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:

    OpenAccountRequest:
      type: object
      required:
        - userId
        - currency
      properties:
        userId:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        currency:
          type: string
          enum: [USD, GBP, EUR]
          example: USD
        name:
          type: string
          example: My Trading Account

    AccountResponse:
      type: object
      required:
        - id
        - userId
        - name
        - balance
        - currency
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        userId:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440001
        name:
          type: string
          example: My Trading Account
        balance:
          type: number
          example: 0.00
        currency:
          type: string
          enum: [USD, GBP, EUR]
          example: USD
        status:
          type: string
          enum: [ACTIVE, SUSPENDED, CLOSED]
          example: ACTIVE
        createdAt:
          type: string
          format: date-time
          example: "2026-06-20T12:00:00Z"

    AccountListResponse:
      type: object
      required:
        - accounts
      properties:
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/AccountResponse'

    ErrorResponse:
      type: object
      required:
        - status
        - error
        - details
      properties:
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Validation failed
        details:
          type: array
          items:
            type: string
          example:
            - "currency: must be one of USD, GBP, EUR"

    TopUpAccountRequest:
      type: object
      required:
        - userId
        - amount
      properties:
        userId:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440001
        amount:
          type: integer
          minimum: 1
          maximum: 10000000
          example: 1000

    TopUpAccountResponse:
      type: object
      required:
        - accountId
        - newBalance
        - currency
        - ledgerEntryId
        - timestamp
      properties:
        accountId:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        newBalance:
          type: number
          example: 1100.00
        currency:
          type: string
          enum: [USD, GBP, EUR]
          example: USD
        ledgerEntryId:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440002
        timestamp:
          type: string
          format: date-time
          example: "2026-06-24T12:00:00Z"

    TransactionListResponse:
      type: object
      required:
        - transactions
        - page
        - totalPages
        - totalCount
      properties:
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/TransactionResponse'
        page:
          type: integer
        totalPages:
          type: integer
        totalCount:
          type: integer

    TransactionResponse:
      type: object
      required:
        - id
        - type
        - assetType
        - amount
        - currency
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [CREDIT, DEBIT]
        assetType:
          type: string
          enum: [CASH, STOCK_BUY, STOCK_SELL]
        amount:
          type: number
        currency:
          type: string
        ticker:
          type: string
          nullable: true
        shares:
          type: number
          nullable: true
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time